building liveClass  and classWork routes

+++ routes/class/liveClass/liveClasses.js
@@ -6,12 +6,12 @@ const connectionPromise=require('../../../database & models/databaseConnection')
 // the route for staff to be able to set live classes for their students
 
 router.post('/',Authenticate,async(req,res)=>{
-    const{courseId,classGroupId,topic,platform,link}=req.body;
+    const{classId,topic,platform,link,classTime}=req.body;
     const userId=req.user.Id;
     try{
         await connectionPromise.query(
-            `INSERT INTO liveclasses (CourseId, ClassGroupId, LecturerId, Topic, Platform, Link) VALUES (?, ?, ?, ?, ?, ?)`,
-            [courseId, classGroupId, userId, topic, platform, link]
+            `INSERT INTO liveclasses (ClassId, LecturerId, Topic, Platform, Link,ClassTime) VALUES (?, ?, ?, ?, ?, ?)`,
+            [classId, userId, topic, platform, link,classTime]
           );
           res.status(201).json('Live class had been created successfully...');
     }catch{(err)=>{
@@ -20,12 +20,18 @@ router.post('/',Authenticate,async(req,res)=>{
 })
 router.delete('/',Authenticate,async(req,res)=>{
     const{classId}=req.body.liveClassId;
-    try{
-        await connectionPromise.query(`delete from liveclasses where Id=?`,[classId]);
-        res.status(200).json({"response":'Live class had been deleted successfully...'})
+    const userRole=req.user.role
+    if(userRole=='staff'){
+        try{
+            await connectionPromise.query(`delete from liveclasses where Id=?`,[classId]);
+            res.status(200).json({"response":'Live class had been deleted successfully...'})
+        }
+        catch{
+            res.status(500).json({"error":'Error deleting live class...'})
+        }
     }
-    catch{
-        res.status(500).json({"error":'Error deleting live class...'})
+    else{
+        res.status(403).json({"error":'You are not authorized to delete live classes...'})
     }
 })
 //route to get live classes posted
@@ -38,8 +44,10 @@ router.get('/',Authenticate,async(req,res)=>{
             res.status(200).json(liveClasses[0]);
         }
         else if(userRole=='student'){
-            const courses=await connectionPromise.query(`select CourseGroup from registration where StudentId=?`[userId])
-            res.status(200).send(courses[0])
+            const [classGroups]=await connectionPromise.query('select ClassId from  roommembership where MemberId=?',[userId])
+            const groupIds=classGroups.map(id=>id.ClassId)
+            const [courses]=await connectionPromise.query(`select * from liveclasses where ClassId IN (${groupIds})`)
+            res.status(200).send(courses)
         
         }
     }catch{(err)=>{
